requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.1
django-autoslug==1.9.3
django-toolbelt==0.0.1
djangorestframework==3.6.3
gunicorn==19.7.1
oauth2client==4.1.0
packaging==16.8
psycopg2==2.7.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
Pygments==2.2.0
pyparsing==2.2.0
pytz==2017.2
requests==2.14.2
requests-futures==0.9.7
rsa==3.4.2
six==1.10.0
static3==0.7.0
uritemplate==3.0.0
